backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
aiofiles
supabase
rank-bm25